homework-1/pyproject.toml
[project]
name = "pattern-2025"
version = "0.1.0"
description = "pattern 2025 assignment"
readme = "README.md"
requires-python = "==3.10.0"
dependencies = [
    "isort>=7.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "scipy>=1.15.3",
    "tqdm>=4.67.1",
]
